+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Core
 - Handler priority system with `utils.Priority` decorator
 - New event (`utils.Events.NODE_FINALIZE`) for doing work after all handlers called and node traversed
+- Instead of reportme, reports are now in JSON format ([#2](https://github.com/thg-consulting/inspectortiger/issues/2))
 ### Plugins
 #### Context
 - Context management improved with scope finalizing